pyproject.toml
[project]
name = "rf-swarm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "cupy-cuda12x>=13.6.0",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "scipy>=1.16.1",
    "tyro>=0.9.28",
]
